Implement the Highest Common Factor
Highest Common Factor also known as the greatest common divisor (GCD)
of two or more integers, which are not all zero, is the largest positive
integer that divides each of the integers.

This solution makes use of the Euclidean Algorithm, a variant in which
the difference of the numbers a and b is replaced by the remainder of the
Euclidean divison (also called division with remainder) of a by b.